requirements.txt
# ============================================================================
# iaq4j — ML platform for indoor air quality prediction
# Python 3.9+ compatible. Python 3.12+ recommended (Docker default).
# KAN sidecar runs Python 3.9 (vendored efficient-kan).
# ============================================================================

# ============================================================================
# Core Web Framework
# ============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.6.0
python-multipart>=0.0.6

# ============================================================================
# Deep Learning - PyTorch Stack
# ============================================================================
torch>=2.9.0
torchvision>=0.24.0
torchaudio>=2.9.0

# ============================================================================
# Machine Learning & Data Science
# ============================================================================
scikit-learn>=1.5.0
numpy>=2.0.0
pandas>=2.2.0

# ============================================================================
# Time Series Database
# ============================================================================
influxdb>=5.3.1                # For InfluxDB 1.x (DataFrameClient)
influxdb-client>=1.38.0        # For InfluxDB 2.x (InfluxDBClient)

# ============================================================================
# Visualization
# ============================================================================
matplotlib>=3.8.0
seaborn>=0.13.0

# ============================================================================
# Utilities
# ============================================================================
pyyaml>=6.0.0
requests>=2.31.0
httpx>=0.27.0
tqdm>=4.66.0
joblib>=1.3.0

# ============================================================================
# KAN (Kolmogorov-Arnold Networks)
# ============================================================================
# KAN is vendored in app/kan.py — runs in-process on Python 3.9, or in the
# kan_sidecar Docker container on Python 3.12+. Do NOT pip install efficient-kan.

# ============================================================================
# Field Mapping (Optional — fuzzy matching for CSV column headers)
# ============================================================================
# rapidfuzz>=3.0.0            # Optional: fuzzy field mapping (python -m iaq4j map-fields)

# ============================================================================
# Development Tools
# ============================================================================
pytest>=8.0.0
pytest-cov>=5.0.0
hypothesis>=6.100.0
# ruff                         # Linter + formatter (replaces black + flake8 + isort)
# mypy>=1.8.0                  # Type checker

# ============================================================================
# SETUP
# ============================================================================
#
# python3.12 -m venv venv
# source venv/bin/activate
# pip install --upgrade pip setuptools wheel
# pip install -r requirements.txt
#
# NOTE: Existing trained model artifacts (.pkl scalers) are incompatible with
# numpy 2.x. Retrain all models after upgrading:
#   python -m iaq4j train --model all --epochs 200
#
# ============================================================================
